[P4PU-745] fixing the involuntary session and local storage clear
