+++ src/routes/AuthCallback/index.tsx
@@ -4,6 +4,7 @@ import { ArcRoutes } from '../routes';
 import { TokenResponse } from '../../../generated/data-contracts';
 import { Box, CircularProgress } from '@mui/material';
 import { tokenResponseSchema } from '../../../generated/zod-schema';
+import utils from 'utils';
 
 export default function AuthCallback() {
   const result = useLoaderData() as TokenResponse | number;
@@ -11,7 +12,7 @@ export default function AuthCallback() {
 
   if (checkToken.success) {
     // if we have a formal token
-    window.localStorage.setItem('accessToken', (result as TokenResponse).accessToken);
+    utils.storage.user.setToken((result as TokenResponse).accessToken);
     window.location.replace(ArcRoutes.DASHBOARD);
   } else if (typeof result === 'number') {
     // if we have an error code
