+++ src/utils/interceptors.ts
@@ -1,7 +1,6 @@
 import { NavigateFunction } from 'react-router-dom';
 import utils from 'utils';
 import { Client } from 'models/Client';
-import { sessionClear } from './session';
 import { ArcRoutes } from 'routes/routes';
 import { AxiosError, InternalAxiosRequestConfig } from 'axios';
 
@@ -25,7 +24,8 @@ export const setupInterceptors = (client: Client, navigate: NavigateFunction) =>
     (error) => {
       if (error.response.status === 401) {
         /* This is a placeholder, in this case I think there should be an attempt to refresh the token, or if that's not possible, to redirect at login. */
-        sessionClear(() => navigate(ArcRoutes.COURTESY_PAGE));
+        utils.storage.user.logOut();
+        navigate(ArcRoutes.COURTESY_PAGE);
       }
     }
   );
