--- src/utils/session.ts
@@ -1,5 +1,0 @@
-export const sessionClear = (onSuccess?: () => void) => {
-  window.sessionStorage.clear();
-  window.localStorage.clear();
-  if (onSuccess) onSuccess();
-};
