+++ src/utils/storage.ts
@@ -1,7 +1,8 @@
 import { signal } from '@preact/signals-react';
 
 enum SessionItems {
-  OPTIN = 'OPTIN'
+  OPTIN = 'OPTIN',
+  TOKEN = 'accessToken'
 }
 
 /** set a session item and return his value. If not possible returs null */
@@ -16,6 +17,12 @@ const setSessionItem = (key: SessionItems, value: string) => {
 
 /** get a session item and return his value. If not possible returns null */
 const getSessionItem = (key: SessionItems) => sessionStorage.getItem(key);
+
+/** clear both session and local storage */
+const clear = () => {
+  window.sessionStorage.clear();
+  window.localStorage.clear();
+};
 
 const optin = signal<boolean>(Boolean(getSessionItem(SessionItems.OPTIN)));
 
@@ -33,5 +40,11 @@ export default {
     clear: () => {
       if (setSessionItem(SessionItems.OPTIN, 'false')) optin.value = false;
     }
+  },
+  user: {
+    hasToken: () => Boolean(getSessionItem(SessionItems.TOKEN)),
+    /** clear both session and local storage */
+    logOut: clear,
+    setToken: (token: string) => setSessionItem(SessionItems.TOKEN, token)
   }
 };
